// bookstore-webapp/src/generated/cart-service.ts
export interface components {
    schemas: {
        CartItem: {
            bookId: string;
            quantity: number;
        };
        Cart: {
            id: string;
            items: components["schemas"]["CartItem"][];
            subtotal: number;
            updatedAt?: string;
        };
        Error: {
            code: number;
            message: string;
            description?: string;
            moreInfo?: string;
        };
    };
    responses: never;
    parameters: never;
    requestBodies: never;
    headers: never;
    pathItems: never;
}